requirements.txt
discord.py
gspread
oauth2client
requests
python-dotenv
